feat(omnibar): refactor attachment components for improved functionality and clarity
- Simplified the `AttachMenu` component by integrating file input resolution logic into a new `fileChannels.js` module, enhancing maintainability.
- Updated the `DropdownMenu` to utilize the new file input structure, improving the user experience when attaching files.
- Introduced a `TabPicker` component for selecting recent tabs, enhancing the omnibar's attachment capabilities.
- Refined the `useMentionPicker` hook to return a handled state for key events, improving interaction with the mention picker.

These changes streamline the attachment process and enhance the overall user experience within the omnibar.
