+++ special-pages/pages/new-tab/app/omnibar/components/chat-tools/tab-attachment/fileChannels.js
@@ -1,0 +1,49 @@
+/**
+ * @typedef {typeof import('../../../strings.json')} Strings
+ * @typedef {{ processFiles: (files: File[]) => Promise<void>, disabled: boolean }} ImageChannel
+ * @typedef {{ processFiles: (files: File[]) => Promise<void>, disabled: boolean, mimeTypes: string[] }} FileChannel
+ * @typedef {{ label: string, accept: string, disabled: boolean, onChange: (event: Event) => Promise<void> }} ResolvedFileInput
+ */
+
+const IMAGE_ACCEPT = 'image/jpeg,image/png,image/webp';
+
+/**
+ * Collapses the optional image / file (PDF) channels into the config for the
+ * single hidden `<input type="file">` rendered by {@link AttachMenu}:
+ *   - `label` / `accept` reflect which of `image`/`file` are non-null.
+ *   - `disabled` is true only when every enabled channel is disabled.
+ *   - `onChange` partitions the picked files back to the right channel
+ *     (anything `image/*` → image, everything else → file).
+ *
+ * @param {(key: keyof Strings) => string} t
+ * @param {ImageChannel | null} image - Pass null to omit the image route.
+ * @param {FileChannel | null} file - Pass null to omit the file (PDF) route.
+ * @returns {ResolvedFileInput}
+ */
+export function resolveFileInput(t, image, file) {
+    const label =
+        image && file ? t('omnibar_attachImageOrFileLabel') : image ? t('omnibar_attachImageLabel') : t('omnibar_attachFileLabel');
+    const accept = [image ? IMAGE_ACCEPT : '', file ? file.mimeTypes.join(',') : ''].filter(Boolean).join(',');
+    const disabled = (image?.disabled ?? true) && (file?.disabled ?? true);
+
+    /** @param {Event} event */
+    const onChange = async (event) => {
+        const input = /** @type {HTMLInputElement} */ (event.currentTarget);
+        if (!input.files || input.files.length === 0) return;
+        const all = Array.from(input.files);
+        /** @type {Promise<void>[]} */
+        const tasks = [];
+        if (image) {
+            const images = all.filter((f) => f.type.startsWith('image/'));
+            if (images.length > 0) tasks.push(image.processFiles(images));
+        }
+        if (file) {
+            const others = all.filter((f) => !f.type.startsWith('image/'));
+            if (others.length > 0) tasks.push(file.processFiles(others));
+        }
+        await Promise.all(tasks);
+        input.value = '';
+    };
+
+    return { label, accept, disabled, onChange };
+}
